refactored autocomplete
