+++ src/autocomplete.ts
@@ -1,18 +1,22 @@
-import _ from 'lodash';
+import {clone} from 'lodash';
 import strip from 'strip-ansi';
-import Command from './command';
-import { IVorpal } from './types';
+import {
+  assembleInput, filterData,
+  getCommandNames,
+  getMatch, getMatchData,
+  getMatchObject,
+  handleTabCounts,
+  parseInput, parseMatchSection,
+} from './autocomplete-utils'
+import {
+  AutocompleteConfigCallback,
+  AutocompleteMatch,
+  AutocompleteOptions,
+  IAutocomplete,
+  Input,
+} from './types/autocomplete'
 
-interface Input {
-  raw;
-  prefix;
-  suffix;
-  context;
-  match?: Command;
-  option?;
-}
-
-const autocomplete = {
+const autocomplete: IAutocomplete = {
   /**
    * Handles tabbed autocompletion.
    *
@@ -24,22 +28,22 @@ const autocomplete = {
    * - Supports piping.
    *
    * @param {String} str
+   * @param {Function} cb
    * @return {String} cb
    * @api public
    */
-
-  exec(str, cb) {
+  exec(str: string, cb: AutocompleteConfigCallback): void {
     let input = parseInput(str, this.parent.ui._activePrompt.screen.rl.cursor);
     const commands = getCommandNames(this.parent.commands);
-    const vorpalMatch = getMatch(input.context, commands, { ignoreSlashes: true });
+    const vorpalMatch = getMatch(input.context as string, commands, { ignoreSlashes: true });
     let freezeTabs = false;
 
-    const end = innerStr => {
+    const end = (innerStr: AutocompleteMatch) => {
       const res = handleTabCounts.call(this, innerStr, freezeTabs);
       cb(undefined, res);
     };
 
-    const evaluateTabs = innerInput => {
+    const evaluateTabs = (innerInput: Input) => {
       if (innerInput.context && innerInput.context[innerInput.context.length - 1] === '/') {
         freezeTabs = true;
       }
@@ -53,21 +57,22 @@ const autocomplete = {
     }
 
     input = getMatchObject.call(this, input, commands);
+
     if (input.match) {
       input = parseMatchSection.call(this, input);
-      getMatchData.call(this, input, function(data) {
-        const dataMatch = getMatch(input.context, data);
+      getMatchData.call(this, input, function(data: string[]) {
+        const dataMatch = getMatch(input.context as string, data);
         if (dataMatch) {
           input.context = dataMatch;
           evaluateTabs(input);
           end(assembleInput(input));
-          return;
+        } else {
+          end(filterData(input.context as string, data));
         }
-        end(filterData(input.context, data));
       });
-      return;
+    } else {
+      end(filterData(input.context as string, commands));
     }
-    end(filterData(input.context, commands));
   },
 
   /**
@@ -76,15 +81,15 @@ const autocomplete = {
    *
    * @param {String} str
    * @param {Array} arr
+   * @param {Object} options
    * @return {String}
    * @api private
    */
-
-  match(str, arr, options) {
+  match(str: string, arr: string[], options: AutocompleteOptions): AutocompleteMatch {
     arr = arr || [];
     options = options || {};
     arr.sort();
-    const arrX = _.clone(arr);
+    const arrX = clone(arr);
     let strX = String(str);
 
     let prefix = '';
@@ -96,7 +101,7 @@ const autocomplete = {
       prefix = parts.length > 0 ? prefix + '/' : prefix;
     }
 
-    const matches = [];
+    const matches: string[] = [];
     for (let i = 0; i < arrX.length; i++) {
       if (strip(arrX[i]).slice(0, strX.length) === strX) {
         matches.push(arrX[i]);
@@ -131,335 +136,5 @@ const autocomplete = {
     return prefix + matches[0].substr(0, longestMatchLength);
   },
 };
-
-/**
- * Tracks how many times tab was pressed
- * based on whether the UI changed.
- *
- * @param {String} str
- * @return {String} result
- * @api private
- */
-
-function handleTabCounts(str, freezeTabs) {
-  let result;
-  if (_.isArray(str)) {
-    this._tabCount += 1;
-    if (this._tabCount > 1) {
-      result = str.length === 0 ? undefined : str;
-    }
-  } else {
-    this._tabCount = freezeTabs === true ? this._tabCount + 1 : 0;
-    result = str;
-  }
-  return result;
-}
-
-/**
- * Looks for a potential exact match
- * based on given data.
- *
- * @param {String} ctx
- * @param {Array} data
- * @return {String}
- * @api private
- */
-
-function getMatch(ctx, data, options?) {
-  // Look for a command match, eliminating and then
-  // re-introducing leading spaces.
-  const len = ctx.length;
-  const trimmed = ctx.replace(/^\s+/g, '');
-  let match = autocomplete.match(trimmed, data.slice(), options);
-  if (_.isArray(match)) {
-    return match;
-  }
-  const prefix = new Array(len - trimmed.length + 1).join(' ');
-  // If we get an autocomplete match on a command, finish it.
-  if (match) {
-    // Put the leading spaces back in.
-    match = prefix + match;
-    return match;
-  }
-  return undefined;
-}
-
-/**
- * Takes the input object and assembles
- * the final result to display on the screen.
- *
- * @param {Object} input
- * @return {String}
- * @api private
- */
-
-function assembleInput(input) {
-  if (_.isArray(input.context)) {
-    return input.context;
-  }
-  const result = (input.prefix || '') + (input.context || '') + (input.suffix || '');
-  return strip(result);
-}
-
-/**
- * Reduces an array of possible
- * matches to list based on a given
- * string.
- *
- * @param {String} str
- * @param {Array} data
- * @return {Array}
- * @api private
- */
-
-function filterData(str, data) {
-  data = data || [];
-  let ctx = String(str || '').trim();
-  const slashParts = ctx.split('/');
-  ctx = slashParts.pop();
-  const wordParts = String(ctx)
-    .trim()
-    .split(' ');
-  let res = data.filter(function(item) {
-    return strip(item).slice(0, ctx.length) === ctx;
-  });
-  res = res.map(function(item) {
-    let parts = String(item)
-      .trim()
-      .split(' ');
-    if (parts.length > 1) {
-      parts = parts.slice(wordParts.length);
-      return parts.join(' ');
-    }
-    return item;
-  });
-  return res;
-}
-
-/**
- * Takes the user's current prompt
- * string and breaks it into its
- * integral parts for analysis and
- * modification.
- *
- * @param {String} str
- * @param {Integer} idx
- * @return {Object}
- * @api private
- */
-
-function parseInput(str, idx): Input {
-  const raw = String(str || '');
-  const sliced = raw.slice(0, idx);
-  const sections = sliced.split('|');
-  const prefixParts = sections.slice(0, sections.length - 1) || [];
-  prefixParts.push('');
-  const prefix = prefixParts.join('|');
-  const suffix = getSuffix(raw.slice(idx));
-  const context = sections[sections.length - 1];
-  return {
-    raw,
-    prefix,
-    suffix,
-    context,
-  };
-}
-
-/**
- * Takes the context after a
- * matched command and figures
- * out the applicable context,
- * including assigning its role
- * such as being an option
- * parameter, etc.
- *
- * @param {Object} input
- * @return {Object}
- * @api private
- */
-
-function parseMatchSection(input) {
-  const parts = (input.context || '').split(' ');
-  const last = parts.pop();
-  const beforeLast = strip(parts[parts.length - 1] || '').trim();
-  if (beforeLast.slice(0, 1) === '-') {
-    input.option = beforeLast;
-  }
-  input.context = last;
-  input.prefix = (input.prefix || '') + parts.join(' ') + ' ';
-  return input;
-}
-
-/**
- * Returns a cleaned up version of the
- * remaining text to the right of the cursor.
- *
- * @param {String} suffix
- * @return {String}
- * @api private
- */
-
-function getSuffix(suffix) {
-  suffix = suffix.slice(0, 1) === ' ' ? suffix : suffix.replace(/.+?(?=\s)/, '');
-  suffix = suffix.slice(1, suffix.length);
-  return suffix;
-}
-
-/**
- * Compile all available commands and aliases
- * in alphabetical order.
- *
- * @param {Array} cmds
- * @return {Array}
- * @api private
- */
-
-function getCommandNames(cmds) {
-  let commands = _.map(cmds, '_name');
-  commands = commands.concat.apply(commands, _.map(cmds, '_aliases'));
-  commands.sort();
-  return commands;
-}
-
-/**
- * When we know that we've
- * exceeded a known command, grab
- * on to that command and return it,
- * fixing the overall input context
- * at the same time.
- *
- * @param {Object} input
- * @param {Array} commands
- * @return {Object}
- * @api private
- */
-
-function getMatchObject(this: IVorpal, input, commands) {
-  const len = input.context.length;
-  const trimmed = String(input.context).replace(/^\s+/g, '');
-  let prefix = new Array(len - trimmed.length + 1).join(' ');
-  let match;
-  let suffix;
-  commands.forEach(function(cmd) {
-    const nextChar = trimmed.substr(cmd.length, 1);
-    if (trimmed.substr(0, cmd.length) === cmd && String(cmd).trim() !== '' && nextChar === ' ') {
-      match = cmd;
-      suffix = trimmed.substr(cmd.length);
-      prefix += trimmed.substr(0, cmd.length);
-    }
-  });
-
-  let matchObject = match
-    ? _.find(this.parent.commands, { _name: String(match).trim() })
-    : undefined;
-
-  if (!matchObject) {
-    this.parent.commands.forEach(function(cmd) {
-      if ((cmd._aliases || []).indexOf(String(match).trim()) > -1) {
-        matchObject = cmd;
-      }
-      return;
-    });
-  }
-
-  if (!matchObject) {
-    matchObject = this.parent.commands.find(cmd => !_.isNil(cmd._catch));
-    if (matchObject) {
-      suffix = input.context;
-    }
-  }
-
-  if (!matchObject) {
-    prefix = input.context;
-    suffix = '';
-  }
-
-  if (matchObject) {
-    input.match = matchObject;
-    input.prefix += prefix;
-    input.context = suffix;
-  }
-  return input;
-}
-
-/**
- * Takes a known matched command, and reads
- * the applicable data by calling its autocompletion
- * instructions, whether it is the command's
- * autocompletion or one of its options.
- *
- * @param {Object} input
- * @param {Function} cb
- * @return {Array}
- * @api private
- */
-
-function getMatchData(input: Input, cb) {
-  const string = input.context;
-  const cmd = input.match;
-  const midOption =
-    String(string)
-      .trim()
-      .slice(0, 1) === '-';
-  const afterOption = input.option !== undefined;
-  if (midOption === true && !cmd._allowUnknownOptions) {
-    const results = [];
-    for (let i = 0; i < cmd.options.length; ++i) {
-      const long = cmd.options[i].long;
-      const short = cmd.options[i].short;
-      if (!long && short) {
-        results.push(short);
-      } else if (long) {
-        results.push(long);
-      }
-    }
-    cb(results);
-    return;
-  }
-
-  function handleDataFormat(str, config, callback) {
-    let data = [];
-    if (_.isArray(config)) {
-      data = config;
-    } else if (_.isFunction(config)) {
-      const cbk =
-        config.length < 2
-          ? function() {}
-          : function(resp) {
-              callback(resp || []);
-            };
-      const res = config(str, cbk);
-      if (res && _.isFunction(res.then)) {
-        res
-          .then(function(resp) {
-            callback(resp);
-          })
-          .catch(function(err) {
-            callback(err);
-          });
-      } else if (config.length < 2) {
-        callback(res);
-      }
-      return;
-    }
-    callback(data);
-    return;
-  }
-
-  if (afterOption === true) {
-    const opt = strip(input.option).trim();
-    const match = cmd.options.find(o => o.short === opt || o.long === opt);
-    if (match) {
-      const config = match.autocomplete;
-      handleDataFormat(string, config, cb);
-      return;
-    }
-  }
-
-  let conf = cmd._autocomplete;
-  conf = conf && conf.data ? conf.data : conf;
-  handleDataFormat(string, conf, cb);
-  return;
-}
 
 export default autocomplete;
