+++ src/types/types.ts
@@ -1,5 +1,7 @@
 import { EventEmitter } from 'events';
-import History from './history';
+import History from '../history';
+import Option from '../option'
+import {IAutocompleteConfig} from './autocomplete'
 
 export interface IVorpal extends EventEmitter {
   parent: IVorpal;
@@ -12,19 +14,20 @@ export interface IVorpal extends EventEmitter {
   session: any;
   prompt(options, cb);
   _commandHelp(command);
-  _useDeprecatedAutocompletion: boolean;
   _send(...argz); // TODO interface to change
 }
 
 export interface ICommand extends EventEmitter {
   commands: ICommand[];
-  options;
+  options: Option[];
   parent: IVorpal;
   _name: string;
   _catch: Function;
   _hidden: boolean;
   _help: Function;
   _aliases: string[];
+  _allowUnknownOptions: boolean
+  _autocomplete: IAutocompleteConfig
   option(flags, description, autocomplete?): ICommand;
   action(fn): ICommand;
   use(fn): ICommand;
@@ -32,8 +35,7 @@ export interface ICommand extends EventEmitter {
 
   cancel(fn): ICommand;
   done(fn);
-  autocomplete(obj);
-  autocompletion(param);
+  autocomplete(obj: IAutocompleteConfig);
   init(fn): ICommand;
   delimiter(delimiter);
   types(types);
